feat: Implement P0 improvements — CI/CD, TypeScript strict mode, type consolidation, docs update
## Changes

### 1. GitHub Actions CI/CD Pipeline
- Add .github/workflows/ci.yml
- Runs type-check, lint, test, build on push/PR
- Ensures broken code cannot be merged

### 2. TypeScript Strict Configuration
- Add noUnusedLocals, noUnusedParameters, noImplicitReturns, noFallthroughCasesInSwitch
- Add type-check and lint:fix scripts to package.json
- Fix type errors in middleware and validation tests
- Create types/jest.d.ts for Jest custom matchers

### 3. Type Consolidation (Coding Style Compliance)
- Move AppSettings interface from lib/settings.ts to types/index.ts
- Update imports in components/Settings.tsx and lib/__tests__/settings.test.ts
- Complies with .claude/rules/coding-style.md rule

### 4. CLAUDE.md Documentation Update
- Remove non-existent app/globals.css
- Add missing files: middleware.ts, animations.ts, i18n.tsx, cn.ts, CircularProgress.tsx, jest.d.ts
- Update file counts to match actual implementation

## Verification
- ✅ npm run type-check (0 errors)
- ✅ npm test (153 tests pass)
- ✅ npm run build (success)
